make mgmt-lxc an upstream to the respective meta-build job
this makes mgmt-lxc a transitive upstream to all builds that depend on the
specific container. since mgmt-lxc is blocking on downstreams and builds
are blocking on upstreams this makes mgmt-lxc not execute while builds are
in progress.
has the somewhat unfortunate downside of mgmt-lxc being able to hold up
a batch integration if it gets queued while builds are in progress. still
better than having builds fail en mass because the original container
is running.