git-monitor replaces previous git locking
git-monitor blocks network bound git access to 5 concurrent connections
this is necessary because git.kde enforces a 5 connection limit and will
reject any additional connection attempts. furthermore we don't want to
hammer git.debian into oblivion.

the new git-monitor is written in ruby and uses drb (ruby ipc) to manage
locking with a daemon component. the daemon holds a list of all pids
that currently hold a lock and mutex locks additional incoming requests
until a new slot opens up.
it also adds additional recovery logic if a process terminates without
first releasing its lock. this prevents lock leaking to the point of
ultimate lock starvation from unreleased locks.
there's also additional logging capabilities to have a change of finding out
what went wrong in case issues should arise.
in particular there's also a log_locks function to force logoutput and
cleanup in case one needs to manually intervene. to use this one would
drb connect to the locker and run the method as seen in the actual git
wrapper.

the git wrapper is also written in ruby to use drb to talk to the daemon,
like the previous shell version it only will lock network bound git
commands, which is a bit of a flaky way to go about it, alas the best
option. locking all operations would cause too great a slowdown to be
viable as jenins' git plugin does rather a lot of rev-parse and the like

git-monitor is supposed to be run as the same user as git (i.e. the jenkins
user) and is started through an upstart job that will also make sure
to respawn the monitor should it explode for whatever reason. stdout is
by default logged to /var/log/upstart/git-monitor.log