fix run decider
drop_while drops up to when the block returns false, that's not what we
want, we want to drop everything that doesn't look like a good build

actually this is slightly inefficient the way it was written, so that might
need redoing at some point. for now simply introduce a bool track for
whether we have passed beyond the valid date range and if so next at
the beginning of the block speeding up the deletes

this is still o(n) in all cases, it does however save us running through
the block and doing costly date parsing. supposedly the idea behind
drop_while was to not run through the entire array. alas, drop_while is not
useful because we actually want to drop_from in a way.