start daily promote only between 6 and 21 oclock
since no one fixes things between 3 and 6 UTC there is no point to running
the promoter during that time frame. what's worse, if there is a full
frameworks rebuild the build blocks itself on the promoter apparently.
in particular but not limited to the fact that the promoter right now
needs to compile a ruby extension which apparently pushes these high
load scenarios even further to the edge