monkey patch ubuntu-defaults-image
livecd-rootfs is automatically pulled in by udi to supply default config
behavior for the ubuntu flavors.
this has the terrible downside that it will force installation of all
packages in the relevant task for a flavor. this list includes libraries
which might be incompatible with what we have in the CI making it
impossible to create an iso.

a current example is libksysguard.
it has a -data package and moved from soversion5 to soversion6 between
plasma 5.2 and 5.3. now since the soversion5 would be in the task list
it would get installed and soversion6 would get installed as part of
regular dep resolution for CI. the two are however incompatible because
they both want the same data package, albeit in different versions.

for now we simply remove the task list for kubuntu-desktop (leaving
minimal, standard and kubuntu-live) which is a bit of a clob solution
as we probably should just blacklist individual packages.