flush manually and use stderr where appropriate
excon somehow manages to introduce substantial buffering, unfortunately
I didn't manage to find the code responsible for it. for now simply flush
after puts in an attached block and log to stderr for event logging.

this should allow immediate output to jenkins but isn't much of a long
term solution