refactor build into a containment class that we can use for promotion
containment encapsulates the logic of setting up an ephemeral container
and running stuff on it.

also add test coverage for this. I just hope this works. also these tests
need refreshing for new docker-api releases as the paths are always fixed
to the supported api version which kinda makes sense