further improve merger speed by reducing pushes to 1 per run
only push once all merge instructions are done. run it once with an
array of branches to push. do not push branches that didn't have a change
anyway (selective list rather than blanket push).

this unfortunately brings in a pile of garbage workarounds to actually
support this as apparently I intentionally designed the merger to work
on remotes all the time, which was slight nonesense really.
two distinct problems:
a) branches would get cleaned for each merge which we can't do since some
   branches will get merges more than once (e.g. stable gets archive and
   backports). to workaround this maintain a list of branches we already
   cleaned
b) branches always merged a remote source even when there was a local one.
   this is nice and dandy with atomic pushes but with postponed pushes
   we must merge the local version of most branches e.g. stable needs
   merging into unstable but then unstable needs the local stable rather
   then remote. to workaround always pick the local version over the remote
   which is a bit meh but should be reliable enough with jenkins
   not actually checking out branches and branches we check out are merge
   targets anyway, so they get updated as part of the merge directives