make builds more robust against docker fail
sometimes under heavy load docker fails to create the devicemap it appears
and fails the build for no good reason [1]. to prevent this retry starting
the container two times for now see how that goes.
this potentially should be moved into containment.rb itself as this is
really something that affects everything.

[1] http://kci.pangea.pub/job/vivid_unstable_khangman/29/console