fix LANG setting
oddly enough language-pack-en-base doesn't set the locale correctly as
plasma-framework failed with

/var/lib/jenkins/ci-tooling/lib/qml/module.rb:34:in `split':
   invalid byte sequence in US-ASCII (ArgumentError)

which is of course a sign of the encoding not being set to UTF-8. so,
forcefully write en_US.UTF-8 to LANG in etc/profile and etc/environment
and run update-locale with it. this should force the correct setup
as previously seen in the dockerfile as well