implicitly call cleanup on container::contain
this was a bit meh in the original design where new would call cleanup
bit in fact it should be called as close to creation as possibly, so that
is contain and not new. the notion here is that one might want to run
more than once if the run didn't succeed. run implies contain so between
runs one would have to explicitly cleanup which is silly given that
containment is supposed to make this stuff more convenient