fix deploy_in_container user creation
I am really very stupid and I do very much love rushed friday evening hacks

so lovely. the code was checking for the jenkins user and if it wasn't
found it would add the jenkins *group* and vice versa.
fixed it by checking user when creating user and group for group....

also change the code order because in an attempt to make this easier to
read. now it first checks the etc data and then acts accordingly rather
than checking and acting and checking and acting which ultimately is what
probably caused me to not notice this mistake as the actions don't stick
out very well in the wall of checks.