setup sudo for the jenkins user in the container
in the future we'll want to on-demand elevate privs rather than enter the
containment as root. doing latter has the profound disadvantage that since
docker lacks uidmapping we write files as host root which prevents jenkins
from deleting them outside the container. so to improve this a bit we'd
want to run the majority of containment work as the contained jenkins user
to reduce the likelihood of accidentally writing files with different
ownership.
this is a slightly big undertaking because things like apt.rb have 0
provisioning for usage without privs though