add (untested) image flattening by export|import piping instead of commit
commiting a container essentially adds another AUFS layer to the AUFS stack
that comprises an image. this is essentially the image history.
unfortunately there is a limit to how big this stack can be, with our
massive history (eeach deployment one layer) we now ran into the limit
and everything exploded.
to prevent this for the future we always flatten the image after creation.
this removes the history (which is slightly useless to us anyway) and
makes the build take a bit longer since it essentially pipes the entire
image through the socket into the RVM and then back the otherway around.

unfortunately since there appears to be no serverside squash operation (not
built in anyway) this is the only 100% reliable way to do it for now.