make sure branches are checked out before cleanup
since we always have a master as we run cleanup magic on it we must make
sure that we reset it properly. actual tracking branches are not used by
jenkins so they will not get updated. which means that if the merger was
run on a different branch and then on master, master would always be a
tracking branch but never get updated since the rest (which would take
care of it) was run on the previous branch and not master.

for instance:
-merger triggers on kubuntu_unstable
 things happen
-merger triggers on master
 (present branch is still kubuntu_unstable)
 cleanup is run
   reset (resets kubuntu_unstable! to where master is)
 wants to merge local master
 nothing to merge

what should have happend is
-merger triggers on master
 checksout master
 calls cleanup
   resets master to where origin/master is
 merge actually happens