fix bindings passing in CI::Containment
start() would pass an unmodified Binds: argument which didn't go through
directbindingsarray and as such wasn't the right format when a single
path was passed in.
e.g.
  Containment(binds: '/tmp')
would end up in docker getting
  Binds: ['/tmp']
when it should have been
  Binds: ['/tmp:/tmp']

Removed the override which now lets Container handle this (correctly).

To prevent this moving forward there now also is a test in containment.