revise fallback upgrades to be more generic
when provisioning the dev version of kubuntu it happens all the time that
it is not present on the hub or for arm, so we need a generic solution
for upgrades from previous to new version. to enable this properly
we now employ lists of possible origins from which we may attempt an
upgrade. this for example means xenial (16.04) can be upgraded from wily or
vivid assuming both are defined. if there is no previous (older) version
then no upgrade will be attempted at all