change prune tech again to support archive properly
we now follow the mtime of the archive rather than link its age to log.
this allows us to clear archive/ even if log doesn't exist anymore.

this is how we should move forward on this as we might need to prune
other bits in the future.

the prune function looks well shit now