refactor trapping to cover all !privd containments and chown to jenkins
docker presently doesn't have stable subuid support so everything created
inside the docker as root is in fact owned by root. the only way to change
this is through an explicit chown which we should be doing in containment
since all trivial access goes through there anyway.
non trivial dockerization needs to handle this manually still, although
I don't think we have such a thing really.

corner case are priviliged containers. those are primarily used for imagers
and iso builders and those in fact want to preserve root as owner on the
files as to assure that the images/isos will have correct file ownership.