adapt remove_container to shitty docker api inconsistency
so...

firstly, container's Image property is different from /all+filter to /json
where former shows the name of the image (even if not tagged anymore)
whereas latter has the id.
additionally in v1.21 a new property ImageID was introduced, which probably
was meant to solve this, we therefore prefer this new property when
available.
also, the docker-api gem being smart merges /all and /json which means
Image of the .info hash in older versions is inconsistent rubbish as
it either can be the id (from /json) or name (from /all) depending on
merge order. to workaround this we always use the /json data only instead
of going through the .info hash.

due to this added complexity we now manually need to translate id to
an image instance to get its tags. this additionally has inconsistency as
for images /all will not list RepoTags and /json will, so we need to
refresh the image to get the data merged from /json before we can get a
repo tag.
since images can have more than one tag we simply get the first and check
if it is pangea. due to the strictness of the cleanup anything that
has more than two tags is not pangea anyway and so it won't match anyway