refactor override application into more readable dimensions and prevent nil
originally we wanted to have overrides override upstream_scm not being
desired. BUT the problem is that we have overrides that override an
upstream_scm where there is none (e.g. liberal rules a la applications/*
apply to applications/meta-yolo which is native and has no upstream_scm, so
changing the upstream_scm via overrides would raise errors on writing
to nil)