port ci-tooling tests to run in parallel
to make this happen we need some additional messing about with the
parallel_test binary as parllel_test is apparently meant to be used from
CLI or via builtin tasks for the most part.
we basically have a replacement test for the old ci tooling tests that
now uses parallel_test instead of ruby directly. additionally the helper
now needs to load the testunit reporting as that'd be mighty crafty to
do via rake now.

pangea testing remains sync as that uses docker and I expect unstable tests
if we were to run them across multiple processors