change sourcer behavior to not wiggle the sources into order
previously we would attempt to be very strict with copying packaging/
and source/ and which directories we want from them respectively.
since mobile needs to work with obscenely fucked up shit bzr repos however
we cannot feasibly wiggle anymore as between our present CIs we cover
every possible combination of native/quilt v1/v3 one-repo/multi-repo, this
ultimately means we need to let everything slide or we have to write
hundreds of lines of code just to impose meaning on meaningless constructs
of source vs. packaging.

as of this change a source is composed by:
- copying source/ to the target directory
- stripping target/source/ of unwanted files
- creating an orig tarball of the content
- copying packaging/ to the target directory (this overwrites stuff from
  source/ potentially)
- building the source package. dpkg-source now is responsible to figure
  out whether whether we want native or not and whether we want a tarball
  or not...