manually set the path for the reporter
as it is using git we'd need bundler to sort the load path. BUT bundler
would require a Gemfile & Gemfile.lock t know that it is installed and
which revision to use, since we have neither as we don't have the master
dir in the containers the bundler setup will not actually do anything
and the resolution of the reporter will fail.
I hope this will be solved soonishy by getting reporter fixed up for
test-unit3 as per a pending pull request.