fix bundle usage in rake
bundler can (and apparently mostly is) automatically injected into running
rubies through rubygems' bundler integration hooks. the problem is that
as part of the bundler setup it conveniently alters the environment to
make sure that all rubies forked from the ruby in question will also load
(the same) bundler instance and set it up accordingly.
this however results in an expectation clash as bundler pack/install
internally asserts that 'setup' only happens once but through this
auto-injection it would happen twice.
the way to avoid this is to run bundler through a clean environment