let pangea_dput switch from changes to dsc automatically
aptly internally does not know how to handle orig-less uploads. for such
uploads .changes will not list the orig tarball but dsc will, since aptly
processes the dsc it thinks the orig should be there when in fact it
shouldn't since it wasn't in the .changes it should not be uploaded

this ultimately is a defect in aptly as we should only upload what is in
the .changes. for the time being however simply allow both .changes or
.dsc as arguments and if attempting to open a .changes instead find its
.dsc and use that to establish the list of files for upload. that way
our external behavior is still correct in that we accept a .changes,
internally we then redirect this into something aptly can correctly deal
with though

Also see https://github.com/smira/aptly/issues/370