fail on build errors
previously this would fail later when expecting exactly one .dsc, that's
causing silly output though and is altogether not the intent, as that is
to fail when the build actually fails. while failing on missing artifacts
is related to that it is not quite the same, a build can succed and fail to
generate the expected artifacts or vice versa.