cleanup all possible container states and anchor running on creation
there is a pre-state to running (created) at which point a container has
no actual StartedAt property (well, it has, but it is unix0, so it's
always too old, which is not actually correct). to deal with this we
use created which is a much more reliable indicator of how old the
container is in terms of a cleanup context. since the smallest age is a day
anyway we don't actually care much about whether it has been running or
created for a day - in both cases something is wrong.