implement bin-only build mode to bootstrap new qt
since 5.6 qtbase and qttools have a circular dependency on arch_indep but
not on arch_dep to build documentation. the way this is meant to bootstrap
is to build the necessary bits binary only first and then indep only.
this change implements former, latter is ultimately covered by rebuilds
of both indep and bin, even if a bit wasteful as they need to rebuild
everything