Allow setting a build dir during init
Firefox and Thunderbird builds on CI fail when the volume is mapped to the
disk. This is most likely a bug in aufs and in order to work around it
we want to build those packages in a temp dir instead.

This preserves the old behaviour of building in the the CWD/build dir with
the resulting binaries in CWD/result