use dir relative to workspace
so. problem is we can't forward mount the repo because then the container
chowning would screw with it and make it owned by jenkins. OTOH we can't
avoid the chowning either because the then the workspace is ugh garbage, so
we must instead copy it into the workspace

somewhat sad, but would be solved if and when we do subuid with docker.
maybe. one day.