change wiping logic for projects' sub jobs
- wipe everything but adt all the time
- only wipe adt if the job is a success

this in particular allows the adt job to offer adt-output artifcats but
if the job is not failing they would get dropped. this is unfortunately
not ideal since we don't retro-actively clean up previous artifacts, but
that would require some additional changes to how the sub jobs run (i.e.
they would have to clean their own artifacts which possibly best be done
on a more global scale anyway ==> major job xml refactor)