make cleanup more efficient by dropping keys from local cache
in a cleaner we have a binary and source cache for their keys. to prevent
excess rountrips when cleaning up binaries we drop packages from the cache
as we remove them from remote. this ultimately shrinks the arrays to
only the stuff we have remote thus allowing binary queries to be quicker
about deciding whether they have a source in the repo