do not purge neon-settings from target to allow possible pins to apply
this is somewhat fishy for a reason or another. we only know that
neon-settings=1 will install on top of the root system, but we do not know
if neon-settings=2 does. That is a general problem with the fact that we
only install and upgrade but then never try to upgrade from system packages

the primary motivation for this is that neon-settings can contain apt pins
which may or may not prevent the proposed repo from upgrading. mind you,
this change has marginal impact as target only gets purged IFF install
failed, at which point neon-settings likely isn't installed anyway