disable nci install check root again
the root tech might have been a stupid idea because it can't work as we
do them now. this kinda needs to be a separate container using a slightly
different testing process than we have here.
namely root needs to install and proposed needs to install, target is
neither needed nor relevant. if proposed installs on top of root AND
proposed upgrades on top of target then by extension proposed is equal or
better than target anyway

# Roots are somewhat broken conceptually. Roots would have to be the first
# package installed. But to do that they would have to be before the target,
# now if the target has lesser quality than the proposed then target will
# fail and render the system into a broken state making proposed unable to
# succeeed. I can't seem to find a way to make this work without having two
# distinct tests. Unfortunately :/