fix releaseme deployment (maybe)
git bundles do not work unless set up, we never really solved this (I
thought we did, alas...). ultimately the problem is that we cannot rely on
bundler to install releaseme as it will install in a special path not
searched by rubygems without the bundler rigging to support it.
to bypass the problem clone releaseme and rake install it to install the
gem proper, I think this should also properly replace the previous version

effectively this will render the gem unversioned, which we don't care about
too much. it will be in the default gem path though so it should be
loadable this way