port upstream_scm default branch selection to releasme where applicable
Project.new has a new param 'origin' this origin is pretty much equal to
the concept of a releaseme origin in that it specifies the quality level
of the project at hand (NB: project == job really, awkward naming and all).
The origin is passed and directly influences the upstream_scm's branch
by resolving the upstream_scm.url through releasme and then picking a
releaseme i18n branch according to our origin (e.g.. when working on a
stable origin project we pick the branch associated with stable i18n).

The origin defaults to unstable (which results in *almost* the same
behavior as before), but can be overridden by setting
```
origin: stable
```
in the projects .yml config. This is extracted by the factorizer and then
passed through the factory into Projects.new where it will be passed to
upstream_scm.

The entire thing only becomes active when the upstream_scm branch was not
overridden from the master default AND the url includes kde.org. i.e. all
of this is noop for !kde projects.

Additionally the branch adjustment only happens after the overrides were
applied. This is to detect a change to the default branch but also to
allow overriding the url. i.e. you can override the upstream_scm.url as
per usual and the new url will be used to conduct the lookup through
releaseme. so, ideally all urls should be resolvable at that point.

To prevent any problems for existing tests that would trigger the relevant
code paths the core testcase now also installs a default stub on the
kde_projects.xml file accessed by releaseme to supply reasonable mock data
that have been stripped from the total data set to plasma+frameworks only
(to save space; full file is 1.5M :O)