dont break on dump by changing the way we deal with Description and friends
previously we'd preserver all whitespaces in a multiline field such as
Description. this however meant that we'd preserve trailing newlines.
this caused problems when dumping these fields as they were not consistent
with other fields.
for other fields we needed to forcefully append a \n, for Description this
resulted in a \n\n. more accidental than anything this worked out for all
test cases because the Description was the last field in a binary paragraph
so we even implicitly relied on this fact to properly preserve pargraphs.
that's all nice and dandy as long as Description is the last field, it does
not have to be though, and in those cases this way of doing things would
break the dump as it would now prematurely end the binary package paragraph
at the Description leaving trailing fields behind.

to prevent this from happening the parsing output has been made consistent
even for multiline. the whitespaces inside the multiline continue to be
preserved, so we preserve Description linebreaks for example. once all
fields of the paragraph have been parsed well iter them all and clean up
multilines with an rstrip which will remove the trailing whitespace
characters and since multilines are joined strings this will effectively
only clean up the last line, which previously incorrectly had a trailing
newline character.

e.g.
```
Description: a
 b
Breaks: foobar
```

was previously `{ 'Description' => "a\nb\n", 'Breaks' => 'foobar' }`

it is now `{ 'Description' => "a\nb", 'Breaks' => 'foobar' }`

when dumping deb822 we now properly force all values to be +\n to
suitably line break each field. additionally in control we manually append
+\n to separate the binary paragraphs from one another.