implement a workaround measure for apt-cacher-ng being rubbish
it serves index from-cache when the Content-Length lines up between server
transfer HTTP 206 (partial) and the cache

```
		if(m_bCheckFreshness && pNextData && m_nSizeSeen == mylen && m_nSizeChecked == mylen-1)
```

problem is that file length is a stupid incomplete measure of equality as
this in particular ignores Last-Modified, which would be an infinitely
better way to tell if stuff changed. I am guessing this is used to
be compatible with daft servers that have an incorrect Last-Modified value,
not that this makes sense because even then serving the right index is
the preferable choice over serving a cached index!

anyway, not sure how to fix it in aptcacher yet, so workaround for now.