put the tooling in /tooling
originally we put it into /var/lib/blahblah to be unifrom between host
and container, BUT this hasn't been true in ages if ever.

while $HOME/tooling on the host was the host tooling, which included
ci-tooling, $HOME/tooling in the container was only the ci-tooling!

with the two tooling types merged into one repo this distinction makes no
sense anymore, more so holding on to it means we are having trouble moving
away from ci-tooling as we cannot move stuff out into the main dir until
we fully deploy the entire dir.

this changes implements this.

the entire tree including .bundle and whatnot now gets put into /tooling
and compat links get set up accordingly to keep things rolling.