disable quiet period for adt, it should follow bin immediately
quiet periods make no sense inside the phases, we should really disable
them entirely for all phase parts, but meh

with adt they are actually bad because the period enables other jobs to
swoop in and claim the build slot. this in particular means that we can
have 2*Nslots concurrent workspaces on one slave if job a_bin and b_bin run
then trigger a_adt and b_adt but those are quiet so c_bin and d_bin hop in
and claim the slots. we now have a, b, c, d workspaces on the slave, and
then potentially cause further disk consumption once a_adt and b_adt start
running

I hope that disabling the quiet period prevents this from happening,
although from a scheduling POV _adt and the competing _bin at that
point would be equal so it probably only reduces it from a 100% chance to
a 50% chance. still better.

ultimately I doubt this will be solved properly until workflows become a
thing