detach mgmt_docker from progenitor
we disregard this for the slaves anyway as nci was never hooked up to
block on deployments on mci (which is where the slaves are mangaged). we
haven't seen a single transient failure ever so I am lead to believe
that the time window where we technically have no image to spin a container
from is so marginally small that blocking docker on it is entirely not
worth the time.

move the dep one level up to tooling.