be brave and also do not block tooling on builds
as with bdaea048267a264645396c0d29414d390baa7f72 we technically do this
already for slaves for entirely the same reasons as in that commit.

as a result we'll also try if things will work without this block.
the impact window is technically a bit larger here as we first rm_rf and
then copy, so depending on the amount of crap being copied things could
take a while (unlike image tagging which should be just mutex{x.name=y}