add a new helper to loop for a job loop chain
as jobs wait on upstreams the jobs themselves being upstreams of their
upstreams caues an infinite wait loop from which there is no escape.
unfortunately jenkins does a poor job of telling us where the loop is.

the new script aggregates all jobs that are part of the upstream tree and
runs through them to find a possible looped chain resuling in a job
blockage