mock projects api use via releaseme
for vcs source we actually mock the api endpoint. its only one and the
tests of the sourcer are fairly blackboxy, so best keep it that way.

for upstream_scm we mock the releaseme api itself since overall api use is
limited, should this change we can always opt to mock on a projects api
level later

drop old xml stub