fix deadlock (maybe?)
every so often a socket process gets stuck, gdb suggest it thread1 is in
shutdown but the busy thread is stuck in the event loop's synchronize.

what I think is happening is that we claim the lock in shutdown! and cancel
the sockets at this point there may be shutdown traffic that the
event loop thread wants to process. however reading implies getting the
lock which it now can't get because we hold it in shutdown!.
why that would only happen some of the time is beyond me, assuming that is
the cause the easy solution is to suspend the reading first. so, disable
event looping and join our loop thread. THEN cancel the sockets.
the way I see it when we cancel sockets we do not care about any of the
remaining traffic that may appear on them, so we should not need to
process anything. we'll also close our session shortly after that so
I am hoping net-ssh will simply be reactive and not get another deadlock
at later point :S