change node assignment of QA jobs
I am rather hoping that jenkins implicitly orders things, which is where
the motivation for this comes from.

we'll want persistent-amd64 as first choice, or cloud-amd64 or backwards
compat unqualified amd64 as fallback.

the idea is that these jobs do not need cores or ram, all they need is
bandwith (well and a bit of disk io), so we can run these on persistent
nodes primarily. when we run out of those we can bounce jobs into the cloud
to unblock stuff, but that should be a fallback not the default

(FWIW from observations I'd say that jenkins will pick an existing node
over provisioning a new one, which is cool, but once a node is provisioned
it seems less clear if a preference is in place. since we want clouds to
become idle asap doing everying we can to get jobs not assigned there is
the goal)