daring experiment on container flushing
output from the child is streamed to us in chunks, as such one line which
was one flush in the child can end up being multiple flushes depending
on the amount of chunks it is split into. to prevent this let's try to
let ruby sort this out. stderr would always get flushed right away, stdout
however is buffered a bit.

the problem is that excessive flushing gets old quickly with the amount of
data we are shoveling into jenkins during builds (our cmake is verbose
and all that). more to the point it makes jenkins more busy than it needs
to be.

so, do not explicitly flush stdout on container output, instead let ruby
decide when to flush. the expectation is that this changes nothing much as
stdout would still be consistent within itself and stderr still autoflushed
causing less load while being functionally equivalent

(NB: if this proofs nonesense the reintroduction of flush should get a
 comment explaining why we explicitly flush and why that is useful)