flush + join
when attach returns, flush stdout to ensure it is synced

when cleaning up our stdout thread do not kill it outright. try to join it
for up to 16 seconds and if that fails kill it. given the behavior that
the attach block actually returns when its good and ready the tread should
also end shortly after the container went down making join the method
of choice for waiting on output to be done