standard buffering is way too aggressive, flush on terminal \n
so, as it turns out stdout buffering is super aggressive leading to some
one minute delay between first print and actual flush.
while reduced flushing is all nice and dandy, this is way too excessive
a slow down.

instead of letting ruby decide on its own, flush manually but only if
the chunk we printed ended in a newline. in those cases it is likely that
this was a complete line and thus worth printing.

this gives us a middle ground solution where print();print() will not flush
but puts() will. that is of course assuming docker just spew things
randomly, in which case the notion would still be that \n denotes the end
of line and we only print complete lines.