add logic to snap extender to create stage-debs parts
stage-debs is a special plugin which is intended to break dep chains. it
takes a list of debs to install and an exclusion. the exclusion list
is ripped out of the dep tree of the debs to install, even if this would
technically break the tree.

the intention is that build-packages and stage-packages used in the
snapcraft.yaml will only have the parts of the tree installed that are
not already provided by the kf5 content snap