do not extend buildtimes anymore it breaks horribly
so, the problem with buildtimes is that it screws dep detection.
ordinarily build-packages will install the stuff into the system apt
and then during priming it will lookup the deps.
BUT
when we stage it as a buildtime part we'll either need to select what to
stage (libfoo because we built against libfoo-dev) or not stage anything
and devise another way to solve this.
this is a bit meh but necessary for the time being.

ideally I'd say that we should install all our dev-part packages from kf5
with a locked version. that way we'd have consistent ABI at least