shrink back to down to 8
on 16 core machines 80% of cpu time gets stolen by the hypervisor for other
vms.

which puts us in a tough spot. we can't really use high-cpu drops as
they have only 20G disk (and 24G ram, which means we can't build in-ram).
option may be 32CPU high-cpu drops but those are a bit on the pricy side.
OTOH we could actually build in-ram on those (have 48, could use 25 for
workspace and be fine for just about all builds)

otherwise we'll have to live with kwin taking an hour. which I am not
really willing to do!