[mci] Add a tech to build special packages for mobile
This projects are special in the sense that they needs to be compiled in
two different configurations,

- Generic
- CAF

magic is done in android-headers package, which packs different
android-config.h based on the Build profile selected (caf or generic),
also those specifically built packages needs to be published into
specific repo instead of root.

This multijob have multiple stages,

- Source
- Build generic
- Build CAF
- Publish

In theory Build generic and Build CAF should be built in the same stage
parallel, but read comment in code on why it is not doing as such.