raise errors if a package failed to install
AFAICT we can only ever install packages that actually exist as we get
the actual package list from aptly directly. this means all packages
are generally meant to install, otherwise we can't tell much about the
state of qml or cmake anyway.

it may be equally acceptable to return, but TBH I think the package being
installable is somewhat indicative of large scale problems, so it probably
should be communicated thusly