fix deploy upgrade fixtures to not include excess env vars
when the fixtures were last refreshed they ran after the test_ci_deployer
which (through mgmt::deployer) set ENV['DIST'] to the test name.
previously we did not enforce ENV cleanup across tests which resulted in
the DIST value getting carried from test_ci_deployer into
test_deploy_upgrade_container which promptly recorded the env variable in
the fixture.

this resulted in flaky test runs as the live input would not match
the fixture expectation when the deployer was not ordered to run before the
upgrader.
as per the recently introduced ENV cleanup this turned into a 100% failure.

easily fixed by adjusting the fixture to be what it is meant to be (no
DIST value)