instead of doing a conditional running check kill and rescue errors
`kill!` unlike `kill` does not rescue resulting servererrors. depending
on the docker version a kill on a running container may come back with
an error though (stupid, I know). so, to deal with this simply call kill
instead of kill! as it's a less error prone approach to making sure the
container is dead.
as an additional measure towards the same end call remove with force
so dockerd itself may murder it if our kill didn't do the trick.