force container removal
on armhf for unknown reasons docker can bug out and fail to kill or stop
a container but fail to report this.
as it turns out rm -f also fails but then actually errors out but
ultimately seems to remove the container anyway. so, while we do not
understand the actual issue rm -f seems to at least make sure dangling
bits get murdered as expected