ditch bundler/setup, with our newly revised way of using ruby its useless
previously we had numerous different nonensense ways to manage gems, now
we always install to system as root and force $HOME isolated gems on
a per user basis (i.e. for jenkins itself) this way the load paths should
be lovely to begin with. conversely the bundler/setup require forcefully
mangles the environment making it harder than necessary to work outside the
confines in forked scripts