do not prune logs anymore
the plan is to have our jenkins_archive.rb move entire build dirs.

we'll want to keep the logs really so we can later look at them, which
we currently can't because the pruner deletes them before the archiver
woudl archive them.
seeing as we want archiving instead of removal for logs we'll simply
drop the pruner altogether

(there may be additional meddling necessary later as logs can be fairly
 large and may quickly fill up the data volume)