do not prune the archives anymore
we clean them in our multijob and later move them to our backup mount.
files we do not clean in the multijob are generally worth keeping, so
purning them is utterly counter productive.

in particular openqa artifacts should be kept for a long while.

it may be necessary to look into pruning the backup mount at some point
though