fix imager mounting to mount in a distinct path as read only
previously we mounted the host tooling in the same path as the host tooling
this was done because we had a separation between host and ci/slave tooling
which has long since gone. the old mount locations however is symlinked
between the actual in-image tooling and the location we want to mount
in, which causes weird hickups where we actually appear with our mount
in the origin (that is: /tooling/ci-tooling symlinks to /var/lib/jenkins/
tooling, and we mount into latter which causes /tooling/ci-tooling/. to be
the same as /tooling/ because of said symlink [even though it goes into
the other direction, not sure how that works])

we now mount the host tooling into /tooling-host instead to avoid the
conflict entirely.

also mark the mount read-only (new feature introduced recently in our
tooling)